Pull up PropagateEvalFoldables
